Updates after review.
