+++ source/databricks/calculation_engine/package/calculation/CalculationResults.py
@@ -58,6 +58,7 @@ class WholesaleResultsContainer:
 
 @dataclass
 class BasisDataContainer:
+    # TODO move calculation_id to mwthod (basis_data_writer.py)
     master_basis_data_for_es_per_ga: DataFrame | None = None
     master_basis_data_for_total_ga: DataFrame | None = None
     time_series_quarter_basis_data_for_total_ga: DataFrame | None = None
